+++ tests/test-dynamic-hooks-importerror/hooks/python_gen_project.py
@@ -1,0 +1,6 @@
+import missing_package
+
+
+def pre_gen_project(project_dir, cookiecutter):
+    cookiecutter['name'] = missing_package.name()
+    cookiecutter['uuid'] = '123'