+++ tests/test-dynamic-hooks-runtime-exception/hooks/python_gen_project.py
@@ -1,0 +1,4 @@
+def pre_gen_project(project_dir, cookiecutter):
+    cookiecutter['name'] = 'Archibald'
+    assert False
+    cookiecutter['uuid'] = str.uuid5()