+++ tests/test-dynamic-hooks/hooks/python_gen_project.py
@@ -1,0 +1,6 @@
+from uuid import uuid4
+
+
+def pre_gen_project(project_dir, cookiecutter):
+    cookiecutter['name'] = 'Archibald'
+    cookiecutter['uuid'] = str(uuid4())